Adding configuration option for include/exclude models from admin.
Adds two configuration options. :included and :excluded, both which take an array of Strings. AbstractModel then filters based on these 2 rules when it is choosing which models are available. This is useful to either remove 1 or 2 models from the admin page, or to define the admin page to operate exclusively on a set list of models.

Merb::Slices::config[:merb_admin][:included] = ["Model1", "Model2", "Model3"]

Merb::Slices::config[:merb_admin][:excluded] = ["ModelZ"]